requirements-dev.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.1
django-filter==2.0.0
djangorestframework==3.8.2
gunicorn==19.9.0
Pillow==5.2.0
psycopg2==2.7.5
python-decouple==3.1
pytz==2018.5
static3==0.7.0
